CanMsg+CanCmt: Correct parsing of timestamps around midnight

+++ src/parsers/tv_comments/tv_comment.rs
@@ -4,6 +4,6 @@ use hdf5::{types::VarLenUnicode, H5Type};
 #[repr(C)]
 pub struct CanCmt {
     pub id: u32,
-    pub ts: u32,
+    pub ts: u64,
     pub value: VarLenUnicode,
 }
